Add --webroot argument to doc/ scripts
Previously, the scripts to generate the website would automatically place
output into the folder /local/www/site/htdocs/LaTeXML. This could cause problems on developer machines, where this folder did not always exist.

This commit adds an optional '--webroot' argument to the scripts in the
doc/ folder that allows to pick a specific directory to place website output
in. It is automatically passed between the different scripts. For backwards
compatibility, when omitted it defaults to /local/www/site/htdocs/LaTeXML.