Rewrite tools/makemanifest in Perl
Previously, the tools/makemanifest script was written in bash. This had
the problem, that the `sort' command was behaving different on certain
Linux installations, leading to inconsistent results.

This commit rewrites the tool in pure perl, thus ensuring consistency
between different systems. Furthermore, this commit adds three new
switches to it, --quiet, --save and --diff. These can be used to
suppress output, automatically overwrite the old MANIFEST file, and show
a diff instead of the full new MANIFEST respectively.